Rename service to App usage service and remove reset
