+++ pupil-spa/src/app/login-success/login-success.component.ts
@@ -6,7 +6,7 @@ import { StorageService } from '../services/storage/storage.service';
 import { DeviceService } from '../services/device/device.service';
 import { SpeechService } from '../services/speech/speech.service';
 import { QuestionService } from '../services/question/question.service';
-import { AppCountService } from '../services/app-count/app-count.service';
+import { AppUsageService } from '../services/app-usage/app-usage.service';
 
 @Component({
   selector: 'app-login-success',
@@ -23,7 +23,7 @@ export class LoginSuccessComponent implements OnInit, AfterViewInit, OnDestroy {
               private deviceService: DeviceService,
               private questionService: QuestionService,
               private speechService: SpeechService,
-              private appCountService: AppCountService,
+              private appUsageService: AppUsageService,
               private elRef: ElementRef) {
     const pupilData = storageService.getItem('pupil');
     const schoolData = storageService.getItem('school');
@@ -33,7 +33,7 @@ export class LoginSuccessComponent implements OnInit, AfterViewInit, OnDestroy {
     this.pupil.dob = pupilData.dob;
     this.school = new School;
     this.school.name = schoolData.name;
-    this.appCountService.increment();
+    this.appUsageService.increment();
   }
 
   async ngOnInit() {
